release v0.0.1

+++ start.sh
@@ -1,9 +1,22 @@
 #!/usr/bin/env bash
 
-#nginx -p /data/workspace/lor/orange_tmp -c /data/workspace/lor/orange/conf/nginx.conf
-pid=`cat /data/workspace/lor/orange/logs/nginx.pid`
-echo "kill "$pid
-kill -s QUIT $pid
+current_path=`pwd`
+pid=""
 
-#ORANGE_CONF=orange/orange.conf && nginx -p `pwd` -c ./conf/nginx.conf
+if [ -f $current_path/logs/nginx.pid ]
+then
+	pid=`cat $current_path/logs/nginx.pid`
+fi
+
+if [ "$pid" = "" ]
+then
+	echo "start orange.."
+else
+	echo "kill "$pid
+	kill -s QUIT $pid
+	echo "restart orange.."
+	
+fi
+
+mkdir -p logs
 nginx -p `pwd` -c ./conf/nginx.conf
